// amamenta-api/src/modules/donator/enums/donatorForm.enum.ts
export enum DonatorReceptor {
  UTIN = "UTIN",
  UCINCO = "UCINCO",
  CRISTO_REI = "CRISTO_REI",
  OTHER = "OTHER",
}

export enum DonatorGuidanceSource {
  HMDR = "HMDR",
  USF = "USF",
  MEDIA = "MEDIA",
  OTHER = "OTHER",
}

export enum SubstanceUseClassification {
  ABUSE = "ABUSE",
  NONE = "NONE",
}